refactor EventsAdapter to separate views
