+++ app/src/main/java/com/github/houseorganizer/houseorganizer/calendar/UpcomingAdapter.java
@@ -1,16 +1,12 @@
 package com.github.houseorganizer.houseorganizer.calendar;
 
-import static com.github.houseorganizer.houseorganizer.calendar.Calendar.CalendarView.MONTHLY;
-import static com.github.houseorganizer.houseorganizer.calendar.Calendar.CalendarView.UPCOMING;
 import static com.github.houseorganizer.houseorganizer.calendar.Calendar.Event.putEventStringsInData;
 import static com.github.houseorganizer.houseorganizer.calendar.UpcomingRowItem.DELIMITER;
 import static com.github.houseorganizer.houseorganizer.calendar.UpcomingRowItem.EVENT;
-import static com.github.houseorganizer.houseorganizer.util.Util.logAndToast;
 
 import android.annotation.SuppressLint;
 import android.app.AlertDialog;
 import android.content.Context;
-import android.net.Uri;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -28,18 +24,10 @@ import androidx.recyclerview.widget.RecyclerView;
 import com.github.houseorganizer.houseorganizer.R;
 import com.github.houseorganizer.houseorganizer.calendar.Calendar.Event;
 import com.github.houseorganizer.houseorganizer.image.ImageHelper;
-import com.google.android.gms.tasks.Task;
-import com.google.firebase.firestore.DocumentReference;
-import com.google.firebase.firestore.FirebaseFirestore;
-import com.google.firebase.firestore.QueryDocumentSnapshot;
 import com.google.firebase.firestore.SetOptions;
-import com.google.firebase.storage.FirebaseStorage;
 import com.google.firebase.storage.StorageReference;
 
-import java.time.LocalDate;
 import java.time.LocalDateTime;
-import java.time.YearMonth;
-import java.time.ZoneOffset;
 import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -48,28 +36,21 @@ import java.util.Locale;
 import java.util.Map;
 
 
-public class EventsAdapter extends RecyclerView.Adapter<RecyclerView.ViewHolder> {
+public class UpcomingAdapter extends CalendarAdapter {
 
-    private final ActivityResultLauncher<String> getPicture;
-
-    private final Calendar calendar;
-    private String eventToAttach;
-    private final FirebaseStorage storage = FirebaseStorage.getInstance();
-    private final FirebaseFirestore db = FirebaseFirestore.getInstance();
     private ArrayList<UpcomingRowItem> items;
 
-    public EventsAdapter(Calendar calendar, ActivityResultLauncher<String> getPicture) {
-        this.calendar = calendar;
-        this.getPicture = getPicture;
-        if (calendar.getView() == UPCOMING) {
-            generateItems(calendar.getEvents());
-        }
-        else {
-            items = new ArrayList<>(calendar.getEvents());
-        }
+    public UpcomingAdapter(Calendar calendar, ActivityResultLauncher<String> getPicture) {
+        super(calendar, getPicture);
     }
 
-    private void generateItems(List<Event> events) {
+    @Override
+    public CalendarAdapter switchView() {
+        return new MonthlyAdapter(calendar, getPicture);
+    }
+
+    @Override
+    void generateItems(List<Event> events) {
         ArrayList<UpcomingRowItem> ret = new ArrayList<>();
         if (events.isEmpty()) {
             items = ret;
@@ -91,21 +72,15 @@ public class EventsAdapter extends RecyclerView.Adapter<RecyclerView.ViewHolder>
         items = ret;
     }
 
-    public class EventViewHolder extends RecyclerView.ViewHolder {
+    public static class EventViewHolder extends RecyclerView.ViewHolder {
         public Button titleView;
-        public TextView descView = null;
-        public ImageButton attachView = null;
+        public TextView descView;
+        public ImageButton attachView;
         public EventViewHolder(View eventView) {
             super(eventView);
-
-            if (calendar.getView() == MONTHLY) {
-                titleView = eventView.findViewById(R.id.event_monthly_title);
-            }
-            else {
-                titleView = eventView.findViewById(R.id.event_upcoming_title);
-                descView = eventView.findViewById(R.id.event_upcoming_date);
-                attachView = eventView.findViewById(R.id.event_upcoming_attach);
-            }
+            titleView = eventView.findViewById(R.id.event_upcoming_title);
+            descView = eventView.findViewById(R.id.event_upcoming_date);
+            attachView = eventView.findViewById(R.id.event_upcoming_attach);
         }
     }
 
@@ -119,10 +94,7 @@ public class EventsAdapter extends RecyclerView.Adapter<RecyclerView.ViewHolder>
 
     @Override
     public int getItemViewType(int position) {
-        if (calendar.getView() == UPCOMING) {
-            return items.get(position).getType();
-        }
-        return 0;
+        return items.get(position).getType();
     }
 
     @NonNull
@@ -131,11 +103,7 @@ public class EventsAdapter extends RecyclerView.Adapter<RecyclerView.ViewHolder>
         Context context = parent.getContext();
         LayoutInflater inflater = LayoutInflater.from(context);
         if (viewType == EVENT) {
-            if (calendar.getView() == MONTHLY) {
-                return new EventViewHolder(inflater.inflate(R.layout.calendar_monthly_cell, parent, false));
-            } else {
-                return new EventViewHolder(inflater.inflate(R.layout.calendar_upcoming_row, parent, false));
-            }
+            return new EventViewHolder(inflater.inflate(R.layout.calendar_upcoming_row, parent, false));
         }
         else {
             return new DelimiterViewHolder(inflater.inflate(R.layout.calendar_upcoming_delimiter, parent, false));
@@ -145,32 +113,16 @@ public class EventsAdapter extends RecyclerView.Adapter<RecyclerView.ViewHolder>
     @Override
     public void onBindViewHolder(@NonNull RecyclerView.ViewHolder holder, int position) {
         if (getItemViewType(position) == EVENT) {
-            if (calendar.getView() == MONTHLY) {
-                prepareMonthlyView((EventViewHolder) holder, position);
-            } else {
-                prepareUpcomingView((EventViewHolder) holder, position);
-            }
+            prepareUpcomingView((EventViewHolder) holder, position);
         }
         else {
             prepareDelimiter((DelimiterViewHolder) holder, position);
         }
     }
 
-    public Calendar getCalendar() {
-        return calendar;
-    }
-
     private void prepareDelimiter(DelimiterViewHolder holder, int position) {
         Delimiter delimiter = ((Delimiter) items.get(position));
         holder.dayView.setText(delimiter.getDate().format(DateTimeFormatter.ofPattern("EEEE dd/MM/yyyy")));
-
-    }
-
-    private void prepareMonthlyView(EventViewHolder holder, int position) {
-        holder.titleView.setText(String.format(Locale.ENGLISH, "%d", position + 1));
-        holder.titleView.setOnClickListener(v -> new AlertDialog.Builder(v.getContext())
-                .setTitle(Integer.toString(position + 1)).setMessage("List of events for this day somehow")
-                .setMessage("List of events for this day somehow").show());
     }
 
     private void prepareUpcomingView(EventViewHolder holder, int position) {
@@ -236,69 +188,7 @@ public class EventsAdapter extends RecyclerView.Adapter<RecyclerView.ViewHolder>
         }
     }
 
-    public void showAddEventDialog(Context ctx, DocumentReference currentHouse, String errMessage) {
-        LayoutInflater inflater = LayoutInflater.from(ctx);
-        final View dialogView = inflater.inflate(R.layout.event_creation, null);
-        new AlertDialog.Builder(ctx)
-                .setTitle(R.string.event_creation_title)
-                .setView(dialogView)
-                .setPositiveButton(R.string.add, (dialog, id) -> {
-                    Task<DocumentReference> pushTask = pushEventFromDialog(dialogView, currentHouse);
-                    if (pushTask != null) {
-                        pushTask.addOnSuccessListener(documentReference -> refreshCalendarView(ctx, currentHouse, errMessage));
-                    }
-                    dialog.dismiss();
-                })
-                .setNegativeButton(R.string.cancel, (dialog, id) -> dialog.dismiss())
-                .show();
-    }
 
-    private Task<DocumentReference> pushEventFromDialog(View dialogView, DocumentReference currentHouse) {
-        Map<String, Object> data = new HashMap<>();
-        final String title = ((EditText) dialogView.findViewById(R.id.new_event_title)).getText().toString();
-        final String desc = ((EditText) dialogView.findViewById(R.id.new_event_desc)).getText().toString();
-        final String date = ((EditText) dialogView.findViewById(R.id.new_event_date)).getText().toString();
-        final String duration = ((EditText) dialogView.findViewById(R.id.new_event_duration)).getText().toString();
-        Map<String, String> event = new HashMap<>();
-        event.put("title", title);
-        event.put("desc", desc);
-        event.put("date", date);
-        event.put("duration", duration);
-        if (Event.putEventStringsInData(event, data)) {
-            return null;
-        }
-        data.put("household", currentHouse);
-        return db.collection("events").add(data);
-    }
-
-    @SuppressLint("NotifyDataSetChanged")
-    public void refreshCalendarView(Context ctx, DocumentReference currentHouse, String errMessage) {
-        db.collection("events")
-                .whereEqualTo("household", currentHouse)
-                .whereGreaterThan("start", LocalDateTime.now().toEpochSecond(ZoneOffset.UTC))
-                .get()
-                .addOnCompleteListener(task -> {
-                    if (task.isSuccessful()) {
-                        ArrayList<Event> newEvents = new ArrayList<>();
-                        for (QueryDocumentSnapshot document : task.getResult()) {
-                            // We assume the stored data is well behaved since it got added in a well behaved manner.
-                            Event event = new Event(
-                                    document.getString("title"),
-                                    document.getString("description"),
-                                    LocalDateTime.ofEpochSecond(document.getLong("start"), 0, ZoneOffset.UTC),
-                                    document.getLong("duration") == null ? 0 : document.getLong("duration"),
-                                    document.getId());
-                            newEvents.add(event);
-                        }
-                        notifyDataSetChanged();
-                        calendar.setEvents(newEvents);
-                        generateItems(newEvents);
-                    } else {
-                        logAndToast(ctx.toString(), errMessage, task.getException(),
-                                ctx, ctx.getString(R.string.refresh_calendar_fail));
-                    }
-                });
-    }
 
     private void removeEventFirestoreAndAdapter(Event event, int position) {
         db.collection("events")
@@ -325,6 +215,7 @@ public class EventsAdapter extends RecyclerView.Adapter<RecyclerView.ViewHolder>
         return retView;
     }
 
+    @SuppressLint("NotifyDataSetChanged")
     private void editEvent(Event eventObj, View dialogView) {
         Map<String, Object> data = new HashMap<>();
         final String title = ((EditText) dialogView.findViewById(R.id.new_event_title)).getText().toString();
@@ -354,21 +245,8 @@ public class EventsAdapter extends RecyclerView.Adapter<RecyclerView.ViewHolder>
         this.notifyDataSetChanged();
     }
 
-    public void pushAttachment(Uri uri) {
-        // Store the image on firebase storage
-        FirebaseStorage storage = FirebaseStorage.getInstance();
-        // this creates the reference to the picture
-        StorageReference imageRef = storage.getReference().child(eventToAttach + ".jpg");
-        imageRef.putFile(uri);
-    }
-
     @Override
     public int getItemCount() {
-        if (calendar.getView() == MONTHLY) {
-            return YearMonth.of(LocalDate.now().getYear(), LocalDate.now().getMonth()).lengthOfMonth();
-        }
-        else {
-            return items.size();
-        }
+        return items.size();
     }
 }
