+++ app/src/main/java/com/github/houseorganizer/houseorganizer/panels/CalendarActivity.java
@@ -11,7 +11,8 @@ import androidx.recyclerview.widget.RecyclerView;
 
 import com.github.houseorganizer.houseorganizer.R;
 import com.github.houseorganizer.houseorganizer.calendar.Calendar;
-import com.github.houseorganizer.houseorganizer.calendar.EventsAdapter;
+import com.github.houseorganizer.houseorganizer.calendar.CalendarAdapter;
+import com.github.houseorganizer.houseorganizer.calendar.UpcomingAdapter;
 import com.google.android.material.bottomnavigation.BottomNavigationView;
 import com.google.firebase.firestore.DocumentReference;
 import com.google.firebase.firestore.FirebaseFirestore;
@@ -19,10 +20,9 @@ import com.google.firebase.firestore.FirebaseFirestore;
 public class CalendarActivity extends AppCompatActivity {
     private DocumentReference currentHouse;
 
-    private EventsAdapter calendarAdapter;
+    private CalendarAdapter calendarAdapter;
     private RecyclerView calendarEvents;
     private final Calendar calendar = new Calendar();
-    private int calendarColumns = 1;
 
     @SuppressLint("NotifyDataSetChanged")
     @Override
@@ -33,12 +33,12 @@ public class CalendarActivity extends AppCompatActivity {
         currentHouse = FirebaseFirestore.getInstance().collection("households").document(getIntent().getStringExtra("house"));
 
         calendarEvents = findViewById(R.id.calendar_screen_calendar);
-        calendarAdapter = new EventsAdapter(calendar,
+        calendarAdapter = new UpcomingAdapter(calendar,
                 registerForActivityResult(new ActivityResultContracts.GetContent(), uri -> calendarAdapter.pushAttachment(uri)));
         calendarEvents.setAdapter(calendarAdapter);
-        calendarEvents.setLayoutManager(new GridLayoutManager(this, calendarColumns));
+        calendarEvents.setLayoutManager(new GridLayoutManager(this, 1));
         calendarAdapter.refreshCalendarView(this, currentHouse, "refreshCalendar:failureToRefresh");
-        findViewById(R.id.calendar_screen_view_change).setOnClickListener(v -> calendarColumns = calendar.rotateCalendarView(this, calendarAdapter, calendarEvents));
+        findViewById(R.id.calendar_screen_view_change).setOnClickListener(v -> calendarAdapter = calendar.rotateCalendarView(this, calendarAdapter, calendarEvents));
         findViewById(R.id.calendar_screen_add_event).setOnClickListener(v -> calendarAdapter.showAddEventDialog(this, currentHouse, "addEvent:failure"));
 
         BottomNavigationView menu = findViewById(R.id.nav_bar);
