+++ app/src/main/java/com/github/houseorganizer/houseorganizer/panels/MainScreenActivity.java
@@ -18,7 +18,7 @@ import androidx.recyclerview.widget.RecyclerView;
 
 import com.github.houseorganizer.houseorganizer.R;
 import com.github.houseorganizer.houseorganizer.calendar.Calendar;
-import com.github.houseorganizer.houseorganizer.calendar.EventsAdapter;
+import com.github.houseorganizer.houseorganizer.calendar.UpcomingAdapter;
 import com.github.houseorganizer.houseorganizer.house.CreateHouseholdActivity;
 import com.github.houseorganizer.houseorganizer.house.HouseSelectionActivity;
 import com.github.houseorganizer.houseorganizer.shop.FirestoreShopList;
@@ -52,7 +52,7 @@ public class MainScreenActivity extends AppCompatActivity {
     private FirebaseFirestore db;
     private FirebaseUser mUser;
     private DocumentReference currentHouse;
-    private EventsAdapter calendarAdapter;
+    private UpcomingAdapter calendarAdapter;
     private RecyclerView calendarEvents;
 
     private TaskList taskList;
@@ -77,7 +77,7 @@ public class MainScreenActivity extends AppCompatActivity {
         loadData();
 
         calendarEvents = findViewById(R.id.calendar);
-        calendarAdapter = new EventsAdapter(calendar,
+        calendarAdapter = new UpcomingAdapter(calendar,
                 registerForActivityResult(new ActivityResultContracts.GetContent(), uri -> calendarAdapter.pushAttachment(uri)));
         calendarEvents.setAdapter(calendarAdapter);
         calendarEvents.setLayoutManager(new GridLayoutManager(this, 1));
